requirements.txt
fastapi==0.109.2
python-dotenv==1.0.1
chrome-aws-lambda==10.1.0
pyppeteer==1.0.2
python-multipart==0.0.6
pydantic==2.6.1
uvicorn==0.27.1
